MDL-62836 Badges: Potential badge recipients by group
This affects the case when a badge is setup with the criteria "manual issue by role".
When searching for a user to award the badge to, the teacher/non-editing teacher can see everyone enrolled on the course, even with SEPARATE GROUPS mode enabled.
With this change, if there are groups on the current course and group mode is enabled, then if the teacher chooses a group, when searching for a user, the results will be filtered according to the group selected.