MDL-62836 Badges: Existing badge recipients by group
When searching for a user to revoke the badge from, the teacher/non-editing teacher can see everyone who was awarded with a badge by himself.
With this change, if there are groups on the current course and group mode is enabled, then if the teacher chooses a group, when searching for a user, the results will be filtered according to the group selected.